feat(otel): a supported WarpTelemetry.clear() for spans, metrics, and the facade (#2248)
* feat(otel): clear() empties a span exporter, retaining its causal context (part of #2208)

Task 5 of the #2208 plan. Both correctness tests pass and each passes alone.

HELD AT THE PLAN'S ESCALATION GATE (Task 5 Step 6). The per-element causal
fold measures QUADRATIC, and at DEFAULT_MAX_SPANS — the buffer this method
must actually be able to clear, five times the plan's 2,000-span gate — it
costs ~5.8 s:

  clear   500 =   33.8 ms
  clear  1000 =   60.4 ms
  clear  2000 =  235.0 ms   (gate was ~200 ms)
  clear  4000 =  929.7 ms
  clear 10000 = 5756.3 ms   (= DEFAULT_MAX_SPANS)

  ratio 2000/1000 = 3.89, 4000/2000 = 3.96  -> quadratic
  one ORSet.piece of the same 2000-set = 1.25 ms -> the fold is 188x the floor

Load average 5.59-8.63 throughout (shared box, an unrelated process pinning a
core), so the absolutes are inflated; the ratios are load-independent and are
what establish the curve.

Not improvised past: the fix is a bulk ORSet.removeAll in :kuilt-crdt, which
the plan forbids this worker from touching.

* feat(otel): clear() empties a metric exporter, local-only by construction (part of #2208)

Task 6 of the #2208 plan. Three tests, each verified passing alone.

The plan's first test asserted only sumValue, while clear() empties five maps
and deletes five keys — so four of the five were unpinned. Widened to populate
and assert every kind. Receipt, under one mutation (histograms.clear() and the
HISTOGRAM_STORE_KEY delete removed):

  widened version  -> RED  (AssertionError)
  plan's version   -> GREEN

Third test pins the KDoc's "never throws" claim, which nothing else covered,
reusing the shared FailDeleteStore rather than a local copy.

aMergeAfterAClearRestoresTheOldSumBecauseAMonotonicJoinCannotForget is
deliberate: a monotonic join has no merge-safe forget.

* feat(otel): WarpTelemetry.clear() fans out across every signal (part of #2208)

Also swaps WarpSpanExporter.clear()'s per-element causal fold for the bulk
ORSet.removeAll landed in #2245, and corrects the plan's two now-false claims.

Re-measured at DEFAULT_MAX_SPANS (10,000), load average 1.43:

  per-element fold  5756.3 ms  (quadratic: 4000/2000 = 3.96)
  ORSet.removeAll      2.3 ms  (4000/2000 = 1.33)

~2500x. Timing harness deleted — it asserts nothing and would be a CI
wall-clock flake.

Task 7's recordId helper uses (id.toLong() shr (8 * i)): Int.shr masks its
operand to 5 bits, so an Int-based version mirrors bytes 0-3 into 4-7 rather
than encoding 64 bits.

* docs(otel): document clear() and correct what it falsifies (part of #2208)

Three of these are staleness rather than additions — each is a claim clear()
falsifies:

- WarpLogRecordExporter's retirement section named windowPass as the only
  driver of retirement; clear() is the second, and the only one that retires
  the whole sealed set at once.
- LogSegmentIndex's key-enumeration sentence now says what that constraint
  costs a CONSUMER, which is the gap #2208 closes.
- The kuilt-primitives skill would not have fired on "reset the telemetry
  store" / "clear my data" — added the phrasing per Task 8 Step 6.

Also fixes the recordId helper in the sibling WarpLogRecordExporterClearTest
(merged in PR 1): Int.shr masks its operand to 5 bits, so the Int form mirrored
bytes 0-3 into 4-7 rather than encoding 64 bits. Injective over Int either way,
so this is legibility, not a bug fix.

* fix(otel): fence WarpMetricExporter's writes against a concurrent clear (part of #2208, part of #2232)

Every write path encoded the whole map under the non-suspend reentrantLock and
persisted OUTSIDE it, with no suspend Mutex holding the encode-write pair
together. A mutation could encode the pre-clear map, suspend, and land those
stale bytes after clear() had deleted the keys.

Pre-clear that race was self-healing — the next mutation re-encodes and
overwrites. After a clear nothing rewrites the key unless that kind is used
afresh, so the resurrection is PERMANENT and a restart brings every pre-clear
series back. clear() is what turned a transient inconsistency into a durable
one, which is why this is fixed here rather than left to its own issue.

Adds writeMutex (coroutine Mutex) around all ten mutators and clear(), with the
encode moved INSIDE it. writeMutex-then-lock ordering, no suspend inside lock.
Mirrors WarpSpanExporter (#1053) and WarpLogRecordExporter (#2187); this was
the third exporter and the last one missing the fence.

Red-then-green, both arms of the new real-threaded test:
  gated arm       expected:<0> but was:<5>   (stale write resurrected the sum)
  stochastic arm  live had card-14, store did not (clear's delete raced a write)
Both green after the mutex.

* docs(otel): the frontier clear is hygiene, not a totality repair (part of #2208)

The KDoc justified clearing the causal frontier by claiming a dangling
predecessor dot would break a totality inferCausalLinks relies on. SpanLink.kt
says the opposite and linkFor bears it out: an unresolved dot returns null and
is logged at debug, never thrown. The invariant does not hold today anyway —
maybeEvict drops spans without touching the frontier.

Behaviour is unchanged and still correct; only the justification is now true.
A link pointing at a deliberately forgotten span is noise, not causality.

* docs(otel): move the clear walkthrough to the whole-journey page (part of #2208)

clear() covers all three signals, but the walkthrough lived only in the Logs
topic — and its load-bearing caveat is about METRICS, which is where a reader
would never have found it. Moved to observability.md, with a short pointer in
each of the three signal topics.

Also re-cut the categories. The old split said "notes and timings" are
forgotten in a way peers respect; that is false for timings — a latency
histogram is a DDSketch over per-replica GCounter buckets, i.e. a metric with
no merge-safe forget, and "timings" is exactly what a reader would call it.
Now: notes and traces are respected; counts, readings, and latency summaries
are local-only.

+++ kuilt-otel/src/commonTest/kotlin/us/tractat/kuilt/otel/WarpLogRecordExporterClearTest.kt
@@ -20,8 +20,10 @@ class WarpLogRecordExporterClearTest {
 
     private val replicaA = ReplicaId("A")
 
+    // Little-endian over a Long: Int.shr masks its operand to 5 bits, so `id shr 32` is `id shr 0`
+    // and an Int-based version would silently mirror bytes 0-3 into 4-7 rather than encode 64 bits.
     private fun recordId(id: Int): ByteString =
-        ByteString(ByteArray(8) { i -> (id shr (8 * i)).toByte() })
+        ByteString(ByteArray(8) { i -> (id.toLong() shr (8 * i)).toByte() })
 
     private fun record(id: Int) = LogRecord(
         recordId = recordId(id),
